Send raw array instead of raw png
